[TT-9908] Rewrite the interview question
This has been rewritten to be a bit more engaging and hopefully prevent
some of the common issues we're seeing with submissions.

There is also a strict .gitignore file to prevent people from working
against this repository if they fork it because we want them creating
new repositories for us to look at.